fix CanvasRectangleCache for scene
